+++ src/main/java/seedu/address/logic/commands/allocate/DeallocateCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.allocate;
 
 import static java.util.Objects.requireNonNull;
 
@@ -6,27 +6,29 @@ import java.util.List;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.event.Event;
 
 /**
- * Deletes a event identified using it's displayed index from the address book.
+ * Frees all employees associated with an event.
  */
-public class DeleteEventCommand extends Command {
+public class DeallocateCommand extends Command {
 
-    public static final String COMMAND_WORD = "delete_ev";
+    public static final String COMMAND_WORD = "free";
 
     public static final String MESSAGE_USAGE = COMMAND_WORD
-            + ": Deletes the event identified by the index number used in the displayed event list.\n"
+            + ": Deallocate the employees identified by the index number used in the displayed event list.\n"
             + "Parameters: INDEX (must be a positive integer)\n"
             + "Example: " + COMMAND_WORD + " 1";
 
-    public static final String MESSAGE_DELETE_EVENT_SUCCESS = "Deleted Event: %1$s";
+    public static final String MESSAGE_FREE_EVENT_SUCCESS = "Deallocate all Employees for: %1$s";
 
     private final Index targetIndex;
 
-    public DeleteEventCommand(Index targetIndex) {
+    public DeallocateCommand(Index targetIndex) {
         this.targetIndex = targetIndex;
     }
 
@@ -39,15 +41,16 @@ public class DeleteEventCommand extends Command {
             throw new CommandException(Messages.MESSAGE_INVALID_EVENT_DISPLAYED_INDEX);
         }
 
-        Event eventToDelete = lastShownList.get(targetIndex.getZeroBased());
-        model.deleteEvent(eventToDelete);
-        return new CommandResult(String.format(MESSAGE_DELETE_EVENT_SUCCESS, eventToDelete));
+        Event eventToFree = lastShownList.get(targetIndex.getZeroBased());
+        eventToFree.freeEmployees();
+        model.setEvent(eventToFree, eventToFree);
+        return new CommandResult(String.format(MESSAGE_FREE_EVENT_SUCCESS, eventToFree.getName()));
     }
 
     @Override
     public boolean equals(Object other) {
         return other == this // short circuit if same object
-                || (other instanceof DeleteEventCommand // instanceof handles nulls
-                && targetIndex.equals(((DeleteEventCommand) other).targetIndex)); // state check
+                || (other instanceof DeallocateCommand // instanceof handles nulls
+                && targetIndex.equals(((DeallocateCommand) other).targetIndex)); // state check
     }
 }
