+++ src/main/java/seedu/address/logic/commands/schedule/DisplayScheduleCommand.java
@@ -1,9 +1,11 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.schedule;
 
 import static java.util.Objects.requireNonNull;
 
 import java.util.List;
 
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.processor.DistinctDatesProcessor;
 import seedu.address.model.Model;
 import seedu.address.model.distinctdate.DistinctDate;
