+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -6,28 +6,28 @@ import static seedu.address.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import seedu.address.logic.commands.AddCommand;
-import seedu.address.logic.commands.AddEventCommand;
-import seedu.address.logic.commands.AssignDateCommand;
-import seedu.address.logic.commands.AutoAllocateCommand;
-import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.DeallocateCommand;
-import seedu.address.logic.commands.DeleteCommand;
-import seedu.address.logic.commands.DeleteEventCommand;
-import seedu.address.logic.commands.DisplayScheduleCommand;
-import seedu.address.logic.commands.DisplayScheduleForDateCommand;
-import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.commands.EditEventCommand;
-import seedu.address.logic.commands.ExitCommand;
-import seedu.address.logic.commands.FetchEventCommand;
-import seedu.address.logic.commands.FindCommand;
-import seedu.address.logic.commands.FindEventCommand;
-import seedu.address.logic.commands.HelpCommand;
-import seedu.address.logic.commands.ListCommand;
-import seedu.address.logic.commands.ListEventCommand;
-import seedu.address.logic.commands.ManualAllocateCommand;
-import seedu.address.logic.commands.StatisticsCommand;
+import seedu.address.logic.commands.allocate.AutoAllocateCommand;
+import seedu.address.logic.commands.allocate.DeallocateCommand;
+import seedu.address.logic.commands.allocate.ManualAllocateCommand;
+import seedu.address.logic.commands.employee.AddCommand;
+import seedu.address.logic.commands.employee.ClearCommand;
+import seedu.address.logic.commands.employee.DeleteCommand;
+import seedu.address.logic.commands.employee.EditCommand;
+import seedu.address.logic.commands.employee.FindCommand;
+import seedu.address.logic.commands.employee.ListCommand;
+import seedu.address.logic.commands.event.AddEventCommand;
+import seedu.address.logic.commands.event.AssignDateCommand;
+import seedu.address.logic.commands.event.DeleteEventCommand;
+import seedu.address.logic.commands.event.EditEventCommand;
+import seedu.address.logic.commands.event.FetchEventCommand;
+import seedu.address.logic.commands.event.FindEventCommand;
+import seedu.address.logic.commands.event.ListEventCommand;
+import seedu.address.logic.commands.general.ExitCommand;
+import seedu.address.logic.commands.general.HelpCommand;
+import seedu.address.logic.commands.schedule.DisplayScheduleCommand;
+import seedu.address.logic.commands.schedule.DisplayScheduleForDateCommand;
+import seedu.address.logic.commands.statistics.StatisticsCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
