+++ src/main/java/seedu/address/logic/parser/AutoAllocateCommandParser.java
@@ -8,7 +8,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 import java.util.Set;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.AutoAllocateCommand;
+import seedu.address.logic.commands.allocate.AutoAllocateCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.tag.Tag;
