+++ src/main/java/seedu/address/logic/parser/FindCommandParser.java
@@ -4,7 +4,7 @@ import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT
 
 import java.util.Arrays;
 
-import seedu.address.logic.commands.FindCommand;
+import seedu.address.logic.commands.employee.FindCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.employee.NameContainsKeywordsPredicate;
