+++ src/main/java/seedu/address/ui/ScheduleBox.java
@@ -10,8 +10,8 @@ import javafx.scene.control.ListCell;
 import javafx.scene.control.ListView;
 import javafx.scene.layout.AnchorPane;
 import seedu.address.logic.Logic;
-import seedu.address.logic.commands.DisplayScheduleForDateCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.commands.schedule.DisplayScheduleForDateCommand;
 import seedu.address.logic.parser.CliSyntax;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.distinctdate.DistinctDate;
