+++ src/test/java/seedu/address/logic/commands/EditEmployeeDescriptorTest.java
@@ -12,7 +12,7 @@ import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.EditCommand.EditEmployeeDescriptor;
+import seedu.address.logic.commands.employee.EditCommand.EditEmployeeDescriptor;
 import seedu.address.testutil.EditPersonDescriptorBuilder;
 
 public class EditEmployeeDescriptorTest {
