+++ src/test/java/seedu/address/logic/parser/EmployeeAddressBookParserTest.java
@@ -13,14 +13,14 @@ import java.util.stream.Collectors;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.ClearCommand;
-import seedu.address.logic.commands.DeleteCommand;
-import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.commands.EditCommand.EditEmployeeDescriptor;
-import seedu.address.logic.commands.ExitCommand;
-import seedu.address.logic.commands.FindCommand;
-import seedu.address.logic.commands.HelpCommand;
-import seedu.address.logic.commands.ListCommand;
+import seedu.address.logic.commands.employee.ClearCommand;
+import seedu.address.logic.commands.employee.DeleteCommand;
+import seedu.address.logic.commands.employee.EditCommand;
+import seedu.address.logic.commands.employee.EditCommand.EditEmployeeDescriptor;
+import seedu.address.logic.commands.employee.FindCommand;
+import seedu.address.logic.commands.employee.ListCommand;
+import seedu.address.logic.commands.general.ExitCommand;
+import seedu.address.logic.commands.general.HelpCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.employee.Employee;
 import seedu.address.model.employee.NameContainsKeywordsPredicate;
