+++ src/test/java/seedu/address/testutil/PersonUtil.java
@@ -8,8 +8,8 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 
 import java.util.Set;
 
-import seedu.address.logic.commands.AddCommand;
-import seedu.address.logic.commands.EditCommand;
+import seedu.address.logic.commands.employee.AddCommand;
+import seedu.address.logic.commands.employee.EditCommand;
 import seedu.address.model.employee.Employee;
 import seedu.address.model.tag.Tag;
